Move createCommonComponents to init-common
