+++ src/js/common/init-common.js
@@ -1,0 +1,11 @@
+import createCommonStore from './store'
+
+export function renderCommonComponents(commonStore) {
+  createLoginWidget(commonStore);
+}
+
+export default function initCommon(reducer) {
+  const commonStore = createCommonStore(reducer);
+  renderCommonComponents(commonStore);
+  return commonStore;
+}
