+++ src/js/edu-benefits/1995/edu-benefits-entry.jsx
@@ -10,11 +10,9 @@ import { Provider } from 'react-redux';
 import initReact from '../../common/init-react';
 import routes from './routes';
 import reducer from './reducer';
-import createCommonStore,  { renderCommonComponents } from '../../common/store';
+import initCommon from '../../common/init-common';
 
-const store = createCommonStore(reducer);
-
-renderCommonComponents(store);
+const store = initCommon(reducer);
 
 const browserHistory = useRouterHistory(createHistory)({
   basename: '/education/apply-for-education-benefits/application/1995'
