+++ src/js/no-react-entry.js
@@ -1,4 +1,4 @@
-import createCommonStore, { renderCommonComponents } from './common/store';
+import initCommon from './common/init-common';
 import createApplicationStatus from './common/components/createApplicationStatus';
 import createEducationApplicationStatus from './edu-benefits/components/createEducationApplicationStatus';
 
@@ -21,8 +21,8 @@ import './legacy/mega-menu.js';
 // New sidebar menu
 import './legacy/sidebar-navigation.js';
 
-const store = createCommonStore();
-renderCommonComponents(store);
+const store = initCommon();
+
 if (pensionPages.has(location.pathname)) {
   createApplicationStatus(store, {
     formId: '21P-527EZ',
